moved core package function
