+++ cx/utilities.go
@@ -5,7 +5,6 @@ import (
 	"github.com/skycoin/cx/cx/ast"
 	"github.com/skycoin/cx/cx/constants"
 	"github.com/skycoin/cx/cx/helper"
-	"os"
 )
 
 // Debug ...
@@ -126,194 +125,12 @@ func (cxprogram *ast.CXProgram) PrintStack() {
 	}
 }
 
-// getFormattedDerefs is an auxiliary function for `GetFormattedName`. This
-// function formats indexing and pointer dereferences associated to `arg`.
-func getFormattedDerefs(arg *ast.CXArgument, includePkg bool) string {
-	name := ""
-	// Checking if we should include `arg`'s package name.
-	if includePkg {
-		name = fmt.Sprintf("%s.%s", arg.Package.Name, arg.Name)
-	} else {
-		name = arg.Name
-	}
-
-	// If it's a literal, just override the name with LITERAL_PLACEHOLDER.
-	if arg.Name == "" {
-		name = constants.LITERAL_PLACEHOLDER
-	}
-
-	// Checking if we got dereferences, e.g. **foo
-	derefLevels := ""
-	if arg.DereferenceLevels > 0 {
-		for c := 0; c < arg.DereferenceLevels; c++ {
-			derefLevels += "*"
-		}
-	}
-	name = derefLevels + name
-
-	// Checking if we have indexing operations, e.g. foo[2][1]
-	for _, idx := range arg.Indexes {
-		// Checking if the value is in data segment.
-		// If this is the case, we can safely display it.
-		idxValue := ""
-		if idx.Offset > ast.PROGRAM.StackSize {
-			// Then it's a literal.
-			idxI32 := helper.Deserialize_i32(ast.PROGRAM.Memory[idx.Offset : idx.Offset+constants.TYPE_POINTER_SIZE])
-			idxValue = fmt.Sprintf("%d", idxI32)
-		} else {
-			// Then let's just print the variable name.
-			idxValue = idx.Name
-		}
-
-		name = fmt.Sprintf("%s[%s]", name, idxValue)
-	}
-
-	return name
-}
-
-// GetFormattedName reads `arg.DereferenceOperations` and builds a string that
-// depicts how an argument is being accessed. Example outputs: "foo[3]",
-// "**bar", "foo.bar[0]". If `includePkg` is `true`, the argument name will
-// include the package name that contains it, such as in "pkg.foo".
-func GetFormattedName(arg *ast.CXArgument, includePkg bool) string {
-	// Getting formatted name which does not include fields.
-	name := getFormattedDerefs(arg, includePkg)
-
-	// Adding as suffixes all the fields.
-	for _, fld := range arg.Fields {
-		name = fmt.Sprintf("%s.%s", name, getFormattedDerefs(fld, includePkg))
-	}
-
-	// Checking if we're referencing `arg`.
-	if arg.PassBy == constants.PASSBY_REFERENCE {
-		name = "&" + name
-	}
-
-	return name
-}
-
-// formatParameters returns a string containing a list of the formatted types of
-// each of `params`, enclosed in parethesis. This function is used only when
-// formatting functions as first-class objects.
-func formatParameters(params []*ast.CXArgument) string {
-	types := "("
-	for i, param := range params {
-		types += GetFormattedType(param)
-		if i != len(params)-1 {
-			types += ", "
-		}
-	}
-	types += ")"
-
-	return types
-}
-
-// GetFormattedType builds a string with the CXGO type representation of `arg`.
-func GetFormattedType(arg *ast.CXArgument) string {
-	typ := ""
-	elt := GetAssignmentElement(arg)
-
-	// this is used to know what arg.Lengths index to use
-	// used for cases like [5]*[3]i32, where we jump to another decl spec
-	arrDeclCount := len(arg.Lengths) - 1
-	// looping declaration specifiers
-	for _, spec := range elt.DeclarationSpecifiers {
-		switch spec {
-		case constants.DECL_POINTER:
-			typ = "*" + typ
-		case constants.DECL_DEREF:
-			typ = typ[1:]
-		case constants.DECL_ARRAY:
-			typ = fmt.Sprintf("[%d]%s", arg.Lengths[arrDeclCount], typ)
-			arrDeclCount--
-		case constants.DECL_SLICE:
-			typ = "[]" + typ
-		case constants.DECL_INDEXING:
-		default:
-			// base type
-			if elt.CustomType != nil {
-				// then it's custom type
-				typ += elt.CustomType.Name
-			} else {
-				// then it's basic type
-				typ += constants.TypeNames[elt.Type]
-
-				// If it's a function, let's add the inputs and outputs.
-				if elt.Type == constants.TYPE_FUNC {
-					if elt.IsLocalDeclaration {
-						// Then it's a local variable, which can be assigned to a
-						// lambda function, for example.
-						typ += formatParameters(elt.Inputs)
-						typ += formatParameters(elt.Outputs)
-					} else {
-						// Then it refers to a named function defined in a package.
-						pkg, err := ast.PROGRAM.GetPackage(arg.Package.Name)
-						if err != nil {
-							println(ast.CompilationError(elt.FileName, elt.FileLine), err.Error())
-							os.Exit(constants.CX_COMPILATION_ERROR)
-						}
-
-						fn, err := pkg.GetFunction(elt.Name)
-						if err == nil {
-							// println(CompilationError(elt.FileName, elt.FileLine), err.ProgramError())
-							// os.Exit(CX_COMPILATION_ERROR)
-							// Adding list of inputs and outputs types.
-							typ += formatParameters(fn.Inputs)
-							typ += formatParameters(fn.Outputs)
-						}
-					}
-				}
-			}
-		}
-	}
-
-	return typ
-}
-
-// SignatureStringOfStruct returns the signature string of a struct.
-func SignatureStringOfStruct(s *ast.CXStruct) string {
-	fields := ""
-	for _, f := range s.Fields {
-		fields += fmt.Sprintf(" %s %s;", f.Name, GetFormattedType(f))
-	}
-
-	return fmt.Sprintf("%s struct {%s }", s.Name, fields)
-}
-
-// IsCorePackage ...
-func IsCorePackage(ident string) bool {
-	for _, core := range constants.CorePackages {
-		if core == ident {
-			return true
-		}
-	}
-	return false
-}
-
 // IsTempVar ...
 func IsTempVar(name string) bool {
 	if len(name) >= len(constants.LOCAL_PREFIX) && name[:len(constants.LOCAL_PREFIX)] == constants.LOCAL_PREFIX {
 		return true
 	}
 	return false
-}
-
-// GetArgSizeFromTypeName ...
-func GetArgSizeFromTypeName(typeName string) int {
-	switch typeName {
-	case "bool", "i8", "ui8":
-		return 1
-	case "i16", "ui16":
-		return 2
-	case "str", "i32", "ui32", "f32", "aff":
-		return 4
-	case "i64", "ui64", "f64":
-		return 8
-	default:
-		return 4
-		// return -1
-		// panic(CX_INTERNAL_ERROR)
-	}
 }
 
 func checkForEscapedChars(str string) []byte {
