+++ cxgo/actions/functions.go
@@ -326,8 +326,8 @@ func processTestExpression(expr *ast.CXExpression) {
 	if expr.Operator != nil {
 		opCode := expr.Operator.OpCode
 		if opCode == constants.OP_ASSERT || opCode == constants.OP_TEST || opCode == constants.OP_PANIC {
-			inp1Type := cxcore.GetFormattedType(expr.Inputs[0])
-			inp2Type := cxcore.GetFormattedType(expr.Inputs[1])
+			inp1Type := ast.GetFormattedType(expr.Inputs[0])
+			inp2Type := ast.GetFormattedType(expr.Inputs[1])
 			if inp1Type != inp2Type {
 				println(ast.CompilationError(CurrentFile, LineNo), fmt.Sprintf("first and second input arguments' types are not equal in '%s' call ('%s' != '%s')", cxcore.OpNames[expr.Operator.OpCode], inp1Type, inp2Type))
 			}
@@ -337,7 +337,7 @@ func processTestExpression(expr *ast.CXExpression) {
 
 // checkIndexType throws an error if the type of `idx` is not `i32` or `i64`.
 func checkIndexType(idx *ast.CXArgument) {
-	typ := cxcore.GetFormattedType(idx)
+	typ := ast.GetFormattedType(idx)
 	if typ != "i32" && typ != "i64" {
 		println(ast.CompilationError(idx.FileName, idx.FileLine), fmt.Sprintf("wrong index type; expected either 'i32' or 'i64', got '%s'", typ))
 	}
@@ -493,8 +493,8 @@ func ProcessGoTos(fn *ast.CXFunction, exprs []*ast.CXExpression) {
 
 func checkMatchParamTypes(expr *ast.CXExpression, expected, received []*ast.CXArgument, isInputs bool) {
 	for i, inp := range expected {
-		expectedType := cxcore.GetFormattedType(expected[i])
-		receivedType := cxcore.GetFormattedType(received[i])
+		expectedType := ast.GetFormattedType(expected[i])
+		receivedType := ast.GetFormattedType(received[i])
 
 		if expr.IsMethodCall && expected[i].IsPointer && i == 0 {
 			// if method receiver is pointer, remove *
@@ -526,8 +526,8 @@ func checkMatchParamTypes(expr *ast.CXExpression, expected, received []*ast.CXAr
 		// These temporary variables' types are not properly being set. That's why we use !cxcore.IsTempVar to
 		// exclude these cases for now.
 		if expr.Operator.OpCode == constants.OP_IDENTITY && !cxcore.IsTempVar(expr.Outputs[0].Name) {
-			inpType := cxcore.GetFormattedType(expr.Inputs[0])
-			outType := cxcore.GetFormattedType(expr.Outputs[0])
+			inpType := ast.GetFormattedType(expr.Inputs[0])
+			outType := ast.GetFormattedType(expr.Outputs[0])
 
 			// We use `isInputs` to only print the error once.
 			// Otherwise we'd print the error twice: once for the input and again for the output
